requirements.txt
minio~=7.2.15
python-dotenv~=1.0.1
spacy~=3.8.5
yake~=0.4.8
numpy~=2.2.5
torch~=2.7.0
nltk~=3.9.1
openai~=1.76.0
transformers~=4.51.3
scikit-learn~=1.6.1
Unidecode~=1.4.0
hf_xet~=1.1.0
keybert~=0.9.0
sentence-transformers~=4.1.0
pymorphy2~=0.9.1
rapidfuzz~=3.13.0
redis~=6.0.0
psycopg2~=2.9.10
pyee~=13.0.0
PyMuPDF~=1.25.5
python-docx~=1.1.2
redis~=6.0.0
